REF: split PriorityQueue into separate module
